adding mkdir working with refactor
